git-svn-id: http://projets.lucterios.org/svn/Clients/SDK@83 69e63612-969a-4371-9866-b14aae95a0cb

+++ build.sh
@@ -1,6 +1,11 @@
 #!/bin/sh
 
 svn checkout http://projets.lucterios.org/svn/Clients/SDK/ . --username=user --password=user123
+if [ $? -ne 0 ]
+then
+	echo "+++ Error checkout +++"
+	exit 1
+fi
 
 rm -f *.zip
 rm -f *.tar.gz
@@ -31,18 +36,33 @@ sed -e "s/\$version_build=X;/\$version_build=$VersionBuild;/" setup.tmp > setup.
 
 ArcFileName=SDK_$VersionMaj-$VersionMin-$VersionRev-$VersionBuild
 
-tar --exclude=.svn -cf $ArcFileName.tar Actions Backup/readme Class CNX/readme Help images Main tutorials ConnectionSDK.inc.php coreIndex.php DeleteModule.php FunctionTool.inc.php GNU_General_Public_License.txt Help.php index.php PathInitial.inc.php ReloadModule.php setup.inc.php
+rm -rf CORE
+svn "export" http://projets.lucterios.org/svn/CORE/ CORE --username=user --password=user123
+if [ $? -ne 0 ]
+then
+	echo "+++ Error Help +++"
+	exit 2
+fi
+
+./phpdoc_generator.sh "."
+if [ $? -ne 0 ]
+then
+	echo "+++ Error Help +++"
+	exit 3
+fi
+
+tar --exclude=.svn -cf $ArcFileName.tar Actions Backup/readme Class CNX/readme Help images Main ConnectionSDK.inc.php coreIndex.php DeleteModule.php FunctionTool.inc.php GNU_General_Public_License.txt Help.php index.php PathInitial.inc.php ReloadModule.php setup.inc.php
 if [ $? -ne 0 ]
 then
 	echo "+++ Error lpk +++"
-	exit 2
+	exit 4
 fi
 cp $ArcFileName.tar SDK
 gzip -f -S .lpk SDK
 if [ $? -ne 0 ]
 then
 	echo "+++ Error lpk +++"
-	exit 2
+	exit 5
 fi
 
 rm -rf bin
@@ -51,4 +71,4 @@ cp SDK.lpk bin/$ArcFileName.lpk
 
 echo "--- Compilation $VersionMaj.$VersionMin.$VersionRev.$VersionBuild Success ---"
 
-cp antBuilderOutput.log Output.txt
+cp antBuilderOutput.log bin/Output.txt
